Prepare for 5.1 filters
Shifts the option-intercept code to a trait to keep the Events_Store class file from getting more unwieldy.

There are no functional changes in this commit, simply reorganization.
